fix: design qa

+++ core/data/api/src/main/kotlin/xyz/duckee/android/core/data/ArtRepository.kt
@@ -27,7 +27,7 @@ interface ArtRepository {
         forSale: Boolean,
         imageUrl: String,
         description: String?,
-        priceInFlow: Int,
+        priceInFlow: Double,
         royaltyFee: Int,
         isImported: Boolean,
         modelName: String,
