+++ core/data/impl/src/main/kotlin/xyz/duckee/android/core/data/ArtRepositoryImpl.kt
@@ -34,7 +34,7 @@ internal class ArtRepositoryImpl @Inject constructor(
         forSale: Boolean,
         imageUrl: String,
         description: String?,
-        priceInFlow: Int,
+        priceInFlow: Double,
         royaltyFee: Int,
         isImported: Boolean,
         modelName: String,
