+++ core/domain/src/main/kotlin/xyz/duckee/android/core/domain/art/UploadArtUseCase.kt
@@ -28,7 +28,7 @@ class UploadArtUseCase @Inject constructor(
         forSale: Boolean,
         imageUrl: String,
         description: String?,
-        priceInFlow: Int,
+        priceInFlow: Double,
         royaltyFee: Int,
         isImported: Boolean,
         modelName: String,
