+++ core/model/src/main/kotlin/xyz/duckee/android/core/model/ArtList.kt
@@ -29,7 +29,7 @@ data class ArtList(
         val imageUrl: String,
         val liked: Boolean,
         val owner: Owner,
-        val priceInFlow: Int,
+        val priceInFlow: Double,
         val royaltyFee: Double,
         val tokenId: Int,
     ) {
