+++ core/network/api/src/main/kotlin/xyz/duckee/android/core/network/model/ResponseArtDetail.kt
@@ -32,7 +32,7 @@ data class ResponseArtDetail(
         val liked: Boolean,
         val owner: Owner,
         val parentToken: Token?,
-        val priceInFlow: Int,
+        val priceInFlow: Double,
         val recipe: Recipe?,
         val royaltyFee: Double,
         val tokenId: Int,
@@ -54,7 +54,7 @@ data class ResponseArtDetail(
             val imageUrl: String,
             val liked: Boolean,
             val owner: Owner,
-            val priceInFlow: Int,
+            val priceInFlow: Double,
             val royaltyFee: Double,
             val tokenId: Int,
         )
